Solved OBD Issues caused by random memory pointers apearing

+++ src/lib/sensors/gear.cpp
@@ -9,7 +9,7 @@ namespace GEAR
     }
   }
 
-  int getGear()
+  float getGear()
   {
     if (!digitalRead(0))
     {
